Don't ignore dependency install scripts in emacs-package-install.
The previous code didn't actually update the script name properly in
the loop where it was trying to install all of an add-on package's
dependencies.  As a result, none of the dependencies' install scripts
were actually invoked.

Thanks to Sébastien Villemot <sebastien@debian.org> for tracking down
the problem, and providing the patch. (closes: #693472)