Invoke each add-on install script correctly as new-style or old-style.
Previously, emacs-package-install would invoke all of the add-on
install scripts in a dependency chain as either old-style or
new-style, based solely on whether or not the package that triggered
the install was old-style or new-style.

Now it should invoke each package's install script based on whether
the package itself is new-style or old-style, as determined by the
presence or absence of the policy-required
/usr/lib/emacsen-common/packages/compat/PACKAGE file.

Thanks to Sébastien Villemot <sebastien@debian.org> for the report.
(closes: #693472)