Require add-on packages to depend on emacsen-common >= 2.0.8.
This should be simpler and safer, and emacsen-common is only ~140k,
which shouldn't be too big a burden.

One specific problem this solves is the handling of
/var/lib/emacsen-common when emacsen-common is purged -- in particular
/var/lib/emacsen-common/state/package/installed/*.  Without the
dependency, emacsen-common can't remove the tree without clobbering
the state for every add-on, but if emacsen-common can't remove it, who
can?

This release changes the following requirements for add-on packages
(see debian-emacs-policy for the details):

  - They must now depend on emacsen-common >= 2.0.8.
  - They don't need to conflict with emacsen-common anymore.
  - They don't need to guard their calls to emacs-install-package.
  - They don't need to guard their calls to emacs-remove-package.
  - They should no longer manage their package/installed/ file directly.

In addition emacsen flavor packages should now depend on
emacsen-common >= 2.0.8.